# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Korb API",
    "description": "Backend API for Korb meal planning application",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "tags": ["users"],
        "summary": "Get Me",
        "operationId": "get_me_api_v1_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": ["users"],
        "summary": "Delete Me",
        "description": "Account deletion stub (App Store compliance).\n\nProduction: call Clerk Backend API.",
        "operationId": "delete_me_api_v1_users_me_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": ["users"],
        "summary": "Update Profile",
        "operationId": "update_profile_api_v1_users_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/products/search": {
      "get": {
        "tags": ["products"],
        "summary": "Search Products",
        "operationId": "search_products_api_v1_products_search_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200,
              "title": "Q"
            }
          },
          {
            "name": "retailers",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 500
                },
                {
                  "type": "null"
                }
              ],
              "title": "Retailers"
            }
          },
          {
            "name": "max_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Price"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponse"
                  },
                  "title": "Response Search Products Api V1 Products Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/compare": {
      "get": {
        "tags": ["products"],
        "summary": "Compare",
        "operationId": "compare_api_v1_products_compare_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ingredient",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200,
              "title": "Ingredient"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/deals": {
      "get": {
        "tags": ["products"],
        "summary": "Deals",
        "operationId": "deals_api_v1_products_deals_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/polls": {
      "post": {
        "tags": ["polls"],
        "summary": "Create Poll",
        "operationId": "create_poll_api_v1_polls_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PollCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PollResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/polls/{poll_id}/vote": {
      "post": {
        "tags": ["polls"],
        "summary": "Vote On Poll",
        "operationId": "vote_on_poll_api_v1_polls__poll_id__vote_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "poll_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Poll Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/polls/active": {
      "get": {
        "tags": ["polls"],
        "summary": "Get Active Polls",
        "operationId": "get_active_polls_api_v1_polls_active_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PollResponse"
                  },
                  "type": "array",
                  "title": "Response Get Active Polls Api V1 Polls Active Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/receipts/scan": {
      "post": {
        "tags": ["receipts"],
        "summary": "Scan Receipt",
        "operationId": "scan_receipt_api_v1_receipts_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiptScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/receipts/auto-refill": {
      "post": {
        "tags": ["receipts"],
        "summary": "Configure Auto Refill",
        "operationId": "configure_auto_refill_api_v1_receipts_auto_refill_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutoRefillRule"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AutoRefillRule": {
        "properties": {
          "ingredient_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Ingredient Name"
          },
          "threshold_days": {
            "type": "integer",
            "maximum": 365.0,
            "minimum": 1.0,
            "title": "Threshold Days",
            "default": 7
          }
        },
        "type": "object",
        "required": ["ingredient_name"],
        "title": "AutoRefillRule"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PollCreate": {
        "properties": {
          "recipe_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recipe Id"
          }
        },
        "type": "object",
        "required": ["recipe_id"],
        "title": "PollCreate"
      },
      "PollResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "recipe_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recipe Id"
          },
          "proposed_by": {
            "type": "string",
            "format": "uuid",
            "title": "Proposed By"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "yes_votes": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Yes Votes",
            "default": []
          },
          "no_votes": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "No Votes",
            "default": []
          }
        },
        "type": "object",
        "required": ["id", "recipe_id", "proposed_by", "is_active"],
        "title": "PollResponse"
      },
      "ProductResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "retailer": {
            "type": "string",
            "title": "Retailer"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "original_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Price"
          },
          "discount_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discount Pct"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "valid_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid To"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "id",
          "retailer",
          "name",
          "description",
          "price",
          "original_price",
          "discount_pct",
          "category",
          "image_url",
          "valid_from",
          "valid_to"
        ],
        "title": "ProductResponse"
      },
      "ProfileUpdate": {
        "properties": {
          "username": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          }
        },
        "type": "object",
        "title": "ProfileUpdate"
      },
      "ReceiptItem": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "price": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Price"
          },
          "quantity": {
            "type": "integer",
            "maximum": 9999.0,
            "minimum": 1.0,
            "title": "Quantity",
            "default": 1
          }
        },
        "type": "object",
        "required": ["name", "price"],
        "title": "ReceiptItem"
      },
      "ReceiptScanRequest": {
        "properties": {
          "retailer": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Retailer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReceiptItem"
            },
            "type": "array",
            "minItems": 1,
            "title": "Items"
          },
          "total": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Total"
          }
        },
        "type": "object",
        "required": ["retailer", "items", "total"],
        "title": "ReceiptScanRequest"
      },
...
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "clerk_id": {
            "type": "string",
            "title": "Clerk Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "household_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Household Id"
          },
          "health_streak_days": {
            "type": "integer",
            "title": "Health Streak Days"
          }
        },
        "type": "object",
        "required": [
          "id",
          "clerk_id",
          "email",
          "username",
          "avatar_url",
          "household_id",
          "health_streak_days"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
      "VoteChoice": {
        "type": "string",
        "enum": ["yes", "no"],
        "title": "VoteChoice"
      },
      "VoteRequest": {
        "properties": {
          "vote": {
            "$ref": "#/components/schemas/VoteChoice"
          }
        },
        "type": "object",
        "required": ["vote"],
        "title": "VoteRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}